refactor: split site config and extract navbar into composable components
- Separate SEO metadata (seo.ts) from lean site config (site.ts)
  following My-Portfolio pattern (SITE_INFO + MAIN_NAV)
- Extract DesktopNav and MobileNav into dedicated components
- Refactor Navbar.tsx as orchestrator (like My-Portfolio's SiteHeader)
- Replace all HeroUI onPress with standard onClick across Chatbot,
  Carousel, pagination buttons (shadcn/ui compatibility)
- Restore disconnected handlers (toggleChat, prevSlide)
- Remove dead HeroUI files (primitives.ts, hero.ts)
- Fix pre-existing TS errors (unused variables)
