+++ src/app/proyectos/components/proyectos-por-area.tsx
@@ -182,7 +182,7 @@ export default function ProyectosPorArea() {
                   size="lg"
                   className="group bg-jci-teal relative overflow-hidden text-white"
                   aria-label="Ver más proyectos"
-                  /* onPress={() => setCurrentPage(currentPage + 1)} */
+                  onClick={() => setCurrentPage(currentPage + 1)}
                 >
                   <span className="bg-jci-black absolute inset-0 h-full w-full -translate-x-full transform transition-transform duration-300 group-hover:translate-x-0"></span>
                   <span className="relative z-10">Ver más</span>
